teacher delete bug

+++ controllers/superuser_staff_controller.js
@@ -50,7 +50,15 @@ module.exports.staffUpdate=async function(req,res){
 
 module.exports.staffDelete=async function(req,res){
     try{
-        let faculty=await UserModel.findByIdAndDelete(req.body.id);
+        let faculty=await UserModel.findById(req.body.id);
+        if(faculty.classSub.length>0){
+            await UserModel.findByIdAndDelete(req.body.id);
+        }
+        else{
+            return res.status(200).json({
+                err: true
+            })
+        }
         return res.status(200).json({
             faculty:faculty   
         });
